+++ threadController/relayController.py
@@ -1,6 +1,11 @@
 import threading
 import logging
 from inbox import Inbox
+import time
+import pymysql
+import configparser
+from threadController.chainController import chainController
+import sys
 
 inbox = Inbox()
 
@@ -9,6 +14,8 @@ class relayController(threading.Thread):
 		self.log = logging.getLogger()
 		self.log.debug("Init relay Controll Thread")
 		threading.Thread.__init__(self)
+		self.config = configparser.RawConfigParser()
+		self.config.read('defaults.cfg')
 
 	def run(self):
 		self.log.debug("Run rely Controll Thread")
@@ -16,6 +23,17 @@ class relayController(threading.Thread):
 
 	@inbox.collate
 	def handle(to, sender, subject, body):
-		log = logging.getLogger()
-		log.info("Incomming Mail")
-		return "550 - Oups"
+		try:
+			log = logging.getLogger()
+			log.info("Incomming Mail")
+			# Seperate in secound thread, than i can replace it faster
+			t = chainController(to, sender, subject, body)
+			t.start()
+			t.join()
+			response = t.retMailInfo()
+			if(response == None):
+				response = "550 - Oups"
+			return response
+		except:
+			log.critical("Something go terrible Wrong: "+str(sys.exc_info()))
+			return "550 Server Error"